Split types and move helpers to utils
