+++ deploy/01_peronio_deploy.ts
@@ -1,7 +1,7 @@
 // deploy/01_peronio_deploy.ts
 import hre from "hardhat";
-import { getConstructorParams } from "../helpers/peronio";
-import { IPeronioConstructorParams } from "../types/utils";
+import { getConstructorParams } from "../utils/helpers";
+import { IPeronioConstructorParams } from "../utils/types/iperonio_constructor_params";
 
 // module.exports = async ({ getNamedAccounts, deployments }) => {
 module.exports = async () => {
