+++ tasks/polygonscan.ts
@@ -2,8 +2,8 @@
 /* eslint-disable @typescript-eslint/no-unsafe-call */
 
 import { task } from "hardhat/config";
-import { getConstructorParams } from "../helpers/peronio";
-import { IPeronioConstructorParams } from "../types/utils";
+import { getConstructorParams } from "../utils/helpers";
+import { IPeronioConstructorParams } from "../utils/types/iperonio_constructor_params";
 
 const peronioConstructor: IPeronioConstructorParams = getConstructorParams();
