+++ utils/types/balance_type.ts
@@ -1,0 +1,9 @@
+import { BigNumber } from "ethers";
+
+export interface BalanceType {
+  pe: BigNumber;
+  usdc: BigNumber;
+  mai: BigNumber;
+  matic: BigNumber;
+  wmatic: BigNumber;
+}
