+++ utils/types/iperonio_constructor_params.ts
@@ -1,0 +1,11 @@
+export interface IPeronioConstructorParams {
+  name: string;
+  symbol: string;
+  usdcAddress: string;
+  maiAddress: string;
+  lpAddress: string;
+  qiAddress: string;
+  quickswapRouterAddress: string;
+  qiFarmAddress: string;
+  qiPoolId: string;
+}
