+++ utils/types/iperonio_initialize_params.ts
@@ -1,0 +1,6 @@
+import { BigNumber } from "ethers";
+
+export interface IPeronioInitializeParams {
+  usdcAmount: BigNumber;
+  startingRatio: number;
+}
